Fixed the SHFILEINFO declaration. Thanks to Andras Kovacs for reporting this.

+++ awt/ShellApi.cs
@@ -316,7 +316,7 @@ namespace awt
         public struct SHFILEINFO
         {
             public IntPtr hIcon;
-            public IntPtr iIcon;
+            public int iIcon;
             public uint dwAttributes;
             [MarshalAs(UnmanagedType.ByValTStr, SizeConst = 260)]
             public string szDisplayName;
